refactor Data.php
