+++ code/Helper/Data.php
@@ -1,330 +1,67 @@
 <?php
 
-require 'algoliasearch.php';
+require_once 'AlgoliaSearch/Version.php';
+require_once 'AlgoliaSearch/AlgoliaException.php';
+require_once 'AlgoliaSearch/ClientContext.php';
+require_once 'AlgoliaSearch/Client.php';
+require_once 'AlgoliaSearch/Index.php';
 
 class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 {
     const BATCH_SIZE           = 100;
     const COLLECTION_PAGE_SIZE = 100;
 
-    const XML_PATH_MINIMAL_QUERY_LENGTH             = 'algoliasearch/ui/minimal_query_length';
-    const XML_PATH_SEARCH_DELAY                     = 'algoliasearch/ui/search_delay';
-
-    const XML_PATH_IS_ALGOLIA_SEARCH_ENABLED        = 'algoliasearch/credentials/is_enabled';
-    const XML_PATH_IS_POPUP_ENABLED                 = 'algoliasearch/credentials/is_popup_enabled';
-    const XML_PATH_APPLICATION_ID                   = 'algoliasearch/credentials/application_id';
-    const XML_PATH_API_KEY                          = 'algoliasearch/credentials/api_key';
-    const XML_PATH_SEARCH_ONLY_API_KEY              = 'algoliasearch/credentials/search_only_api_key';
-    const XML_PATH_INDEX_PREFIX                     = 'algoliasearch/credentials/index_prefix';
-
-    const XML_PATH_IS_INSTANT_ENABLED               = 'algoliasearch/instant/is_instant_enabled';
-    const XML_PATH_REPLACE_CATEGORIES               = 'algoliasearch/instant/replace_categories';
-    const XML_PATH_INSTANT_SELECTOR                 = 'algoliasearch/instant/instant_selector';
-    const XML_PATH_FACETS                           = 'algoliasearch/instant/facets';
-    const XML_PATH_SORTING_INDICES                  = 'algoliasearch/instant/sorts';
-
-    const XML_PATH_PRODUCT_ATTRIBUTES               = 'algoliasearch/products/product_additional_attributes';
-    const XML_PATH_PRODUCT_CUSTOM_RANKING           = 'algoliasearch/products/custom_ranking_product_attributes';
-    const XML_PATH_RESULTS_LIMIT                    = 'algoliasearch/products/results_limit';
-
-    const XML_PATH_CATEGORY_ATTRIBUTES              = 'algoliasearch/categories/category_additional_attributes2';
-    const XML_PATH_INDEX_PRODUCT_COUNT              = 'algoliasearch/categories/index_product_count';
-    const XML_PATH_CATEGORY_CUSTOM_RANKING          = 'algoliasearch/categories/custom_ranking_category_attributes';
-
-    const XML_PATH_EXCLUDED_PAGES                   = 'algoliasearch/pages/excluded_pages';
-
-    const XML_PATH_REMOVE_IF_NO_RESULT              = 'algoliasearch/relevance/remove_words_if_no_result';
-
-    const XML_PATH_NUMBER_OF_PRODUCT_SUGGESTIONS    = 'algoliasearch/ui/number_product_suggestions';
-    const XML_PATH_NUMBER_OF_PRODUCT_RESULTS        = 'algoliasearch/ui/number_product_results';
-    const XML_PATH_NUMBER_OF_CATEGORY_SUGGESTIONS   = 'algoliasearch/ui/number_category_suggestions';
-    const XML_PATH_NUMBER_OF_PAGE_SUGGESTIONS       = 'algoliasearch/ui/number_page_suggestions';
-
-    const XML_PATH_USE_RESULT_CACHE                 = 'algoliasearch/ui/use_result_cache';
-    const XML_PATH_SAVE_LAST_QUERY                  = 'algoliasearch/ui/save_last_query';
-
-    private static $_categoryNames;
-    private static $_activeCategories;
     private static $_rootCategoryId = -1;
-    private static $_categoryAttributes;
-    private static $_productAttributes;
 
-    /**
-     * Predefined Magento product attributes that are used to prepare data for indexing
-     *
-     * @var array
-     */
+
     static private $_predefinedProductAttributes = array('name', 'url_key', 'description', 'image', 'thumbnail');
 
-    /**
-     * Predefined product attributes that will be retrieved from the index
-     *
-     * @var array
-     */
-    static private $_predefinedProductAttributesToRetrieve = array('name', 'url', 'thumbnail_url', 'categories');
+    private $algolia_helper;
 
-    /**
-     * Predefined category attributes that will be retrieved from the index
-     *
-     * @var array
-     */
-    static private $_predefinedCategoryAttributesToRetrieve = array('name', 'url', 'image_url');
+    private $page_helper;
+    private $category_helper;
+    private $product_helper;
 
-    /**
-     * Predefined special attributes
-     *
-     * @var array
-     */
-    static private $_predefinedSpecialAttributes = array('_tags');
-
-    /**
-     * Data prefix to retrieve Algolia search specific data for the entity.
-     *
-     * @var string
-     */
-    private $_dataPrefix = 'algolia_';
+    private $config;
 
     public function __construct()
     {
         \AlgoliaSearch\Version::$custom_value = " Magento (1.1.4)";
+
+        $this->algolia_helper   = new Algolia_Algoliasearch_Helper_Algoliahelper();
+
+        $this->page_helper      = new Algolia_Algoliasearch_Helper_Entity_Pagehelper();
+        $this->category_helper  = new Algolia_Algoliasearch_Helper_Entity_Categoryhelper();
+        $this->product_helper   = new Algolia_Algoliasearch_Helper_Entity_Producthelper();
+
+        $this->config           = new Algolia_Algoliasearch_Helper_Config();
     }
 
-    /**
-     * @param string $name
-     * @return \AlgoliaSearch\Index
-     */
-    public function getIndex($name)
+    public function deleteStoreIndices($storeId = null)
     {
-        return $this->getClient()->initIndex($name);
+        $this->algolia_helper->deleteIndex($this->product_helper->getIndexName($storeId));
+        $this->algolia_helper->deleteIndex($this->category_helper->getIndexName($storeId));
+        $this->algolia_helper->deleteIndex($this->page_helper->getIndexName($storeId));
     }
 
-    public function listIndexes()
+    public function saveConfigurationToAlgolia($storeId = null)
     {
-        return $this->getClient()->listIndexes();
+        $this->algolia_helper->setSettings($this->product_helper->getIndexName($storeId), $this->product_helper->getIndexSettings($storeId));
+        $this->algolia_helper->setSettings($this->category_helper->getIndexName($storeId), $this->category_helper->getIndexSettings($storeId));
+        $this->algolia_helper->setSettings($this->page_helper->getIndexName($storeId), $this->page_helper->getIndexSettings($storeId));
     }
 
-    public function deleteStoreIndex($storeId = NULL)
+    public function getSearchResult($query, $storeId)
     {
-        return $this->getClient()->deleteIndex($this->getIndexName($storeId));
-    }
+        $resultsLimit = $this->config->getResultsLimit($storeId);
 
-    public function query($index_name, $q, $params)
-    {
-        return $this->getClient()->initIndex($index_name)->search($q, $params);
-    }
-
-    public function getIndexName($storeId = NULL)
-    {
-        return (string)$this->getIndexPrefix($storeId) . Mage::app()->getStore($storeId)->getCode();
-    }
-
-    public function mergeSettings($index_name, $settings)
-    {
-        $onlineSettings = array();
-
-        try
-        {
-            $onlineSettings = $this->getIndex($index_name)->getSettings();
-        }
-        catch(\Exception $e)
-        {
-        }
-
-        $removes = array('slaves');
-
-        foreach ($removes as $remove)
-            if (isset($onlineSettings[$remove]))
-                unset($onlineSettings[$remove]);
-
-        foreach ($settings as $key => $value)
-            $onlineSettings[$key] = $value;
-
-        return $onlineSettings;
-    }
-
-    public function setIndexSettings($storeId = NULL)
-    {
-        $index = $this->getIndex($this->getIndexName($storeId).'_products');
-        $index->setSettings($this->getProductIndexSettings($storeId));
-
-        $index = $this->getIndex($this->getIndexName($storeId).'_categories');
-        $index->setSettings($this->getCategoryIndexSettings($storeId));
-
-        $index = $this->getIndex($this->getIndexName($storeId).'_pages');
-        $index->setSettings($this->getPageIndexSettings($storeId));
-    }
-
-    public function getProductIndexSettings($storeId)
-    {
-        $attributesToIndex          = array();
-        $unretrievableAttributes    = array();
-        $attributesForFaceting      = array();
-
-        foreach ($this->getProductAdditionalAttributes($storeId) as $attribute)
-        {
-            if ($attribute['searchable'] == '1')
-            {
-                if ($attribute['order'] == 'ordered')
-                    $attributesToIndex[] = $attribute['attribute'];
-                else
-                    $attributesToIndex[] = 'unordered('.$attribute['attribute'].')';
-            }
-
-            if ($attribute['retrievable'] != '1')
-                $unretrievableAttributes[] = $attribute['attribute'];
-        }
-
-        $customRankings = $this->getProductCustomRanking($storeId);
-
-        $customRankingsArr = array();
-
-        $facets = $this->getFacets();
-
-        foreach($facets as $facet)
-            $attributesForFaceting[] = $facet['attribute'];
-
-        foreach ($customRankings as $ranking)
-            $customRankingsArr[] =  $ranking['order'] . '(' . $ranking['attribute'] . ')';
-
-
-        $indexSettings = array(
-            'attributesToIndex'         => array_values(array_unique($attributesToIndex)),
-            'customRanking'             => $customRankingsArr,
-            'unretrievableAttributes'   => $unretrievableAttributes,
-            'attributesForFaceting'     => $attributesForFaceting,
-        );
-
-        // Additional index settings from event observer
-        $transport = new Varien_Object($indexSettings);
-        Mage::dispatchEvent('algolia_index_settings_prepare', array('store_id' => $storeId, 'index_settings' => $transport));
-        $indexSettings = $transport->getData();
-
-        $mergeSettings = $this->mergeSettings($this->getIndex($this->getIndexName($storeId).'_products'), $indexSettings);
-
-        /**
-         * Handle Slaves
-         */
-
-
-        if (count($this->getSortingIndices()) > 0)
-        {
-            $slaves = array();
-
-            foreach ($this->getSortingIndices() as $values)
-                $slaves[] = $this->getIndexName($storeId).'_products_'.$values['attribute'].'_'.$values['sort'];
-
-            $index = $this->getIndex($this->getIndexName($storeId).'_products');
-            $index->setSettings(array('slaves' => $slaves));
-
-            foreach ($this->getSortingIndices() as $values)
-            {
-                $mergeSettings['ranking'] = array($values['sort'].'('.$values['attribute'].')', 'typo', 'geo', 'words', 'proximity', 'attribute', 'exact', 'custom');
-
-                $index = $this->getIndex($this->getIndexName($storeId).'_products_'.$values['attribute'].'_'.$values['sort']);
-                $index->setSettings($mergeSettings);
-            }
-        }
-
-        unset($mergeSettings['ranking']);
-
-        return $mergeSettings;
-    }
-
-    public function getPageIndexSettings($storeId)
-    {
-        $indexSettings = array(
-            'attributesToIndex'         => array('slug', 'name', 'unordered(content)'),
-        );
-
-        return $indexSettings;
-    }
-
-    public function getCategoryIndexSettings($storeId)
-    {
-        $attributesToIndex          = array();
-        $unretrievableAttributes    = array();
-
-        foreach ($this->getCategoryAdditionalAttributes($storeId) as $attribute)
-        {
-            if ($attribute['searchable'] == '1')
-            {
-                if ($attribute['order'] == 'ordered')
-                    $attributesToIndex[] = $attribute['attribute'];
-                else
-                    $attributesToIndex[] = 'unordered('.$attribute['attribute'].')';
-            }
-
-            if ($attribute['retrievable'] != '1')
-                $unretrievableAttributes[] = $attribute['attribute'];
-        }
-
-        $customRankings = $this->getCategoryCustomRanking($storeId);
-
-        $customRankingsArr = array();
-
-        foreach ($customRankings as $ranking)
-            $customRankingsArr[] =  $ranking['order'] . '(' . $ranking['attribute'] . ')';
-
-        // Default index settings
-        $indexSettings = array(
-            'attributesToIndex'         => array_values(array_unique($attributesToIndex)),
-            'customRanking'             => $customRankingsArr,
-            'unretrievableAttributes'   => $unretrievableAttributes
-        );
-
-        // Additional index settings from event observer
-        $transport = new Varien_Object($indexSettings);
-        Mage::dispatchEvent('algolia_index_settings_prepare', array('store_id' => $storeId, 'index_settings' => $transport));
-        $indexSettings = $transport->getData();
-
-        $this->mergeSettings($this->getIndex($this->getIndexName($storeId).'_categories'), $indexSettings);
-
-        return $indexSettings;
-    }
-
-    private function getClient()
-    {
-        return new \AlgoliaSearch\Client($this->getApplicationID(), $this->getAPIKey());
-    }
-
-    /**
-     * Get array of [product_id => [relevance]
-     *
-     * @param $queryText
-     * @param $storeId
-     * @return array
-     * @throws Exception
-     */
-    public function getSearchResult($queryText, $storeId)
-    {
-        Varien_Profiler::start('Algolia-FullText-getSearchResult');
-
-        if ($this->isInstantEnabled($storeId))
-        {
-            $url = Mage::getBaseUrl().'#q='.$queryText.'&page=0&refinements=%5B%5D&numerics_refinements=%7B%7D&index_name=%22'.$this->getIndexName($storeId).'_products%22';
-
-            header('Location: '.$url);
-
-            die();
-        }
-
-        try
-        {
-            $resultsLimit = $this->getResultsLimit($storeId);
-
-            $answer = $this->query($this->getIndexName($storeId).'_products', $queryText, array(
-                'hitsPerPage' => max(5, min($resultsLimit, 1000)), // retrieve all the hits (hard limit is 1000)
-                'attributesToRetrieve' => 'objectID',
-                'attributesToHighlight' => '',
-                'attributesToSnippet' => '',
-                'removeWordsIfNoResult'=> $this->getRemoveWordsIfNoResult($storeId),
-            ));
-        }
-        catch (Exception $e)
-        {
-            Varien_Profiler::stop('Algolia-FullText-getSearchResult');
-            throw $e;
-        }
+        $answer = $this->algolia_helper->query($this->product_helper->getIndexName($storeId), $query, array(
+            'hitsPerPage' => max(5, min($resultsLimit, 1000)), // retrieve all the hits (hard limit is 1000)
+            'attributesToRetrieve' => 'objectID',
+            'attributesToHighlight' => '',
+            'attributesToSnippet' => '',
+            'removeWordsIfNoResult'=> $this->config->getRemoveWordsIfNoResult($storeId)
+        ));
 
         $data = array();
 
@@ -336,150 +73,14 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
                 $data[$productId] = $resultsLimit - $i;
         }
 
-        Varien_Profiler::stop('Algolia-FullText-getSearchResult');
-
         return $data;
     }
 
-    /**
-     * Return array of all product attributes that can be indexed (all except internal attributes and default attributes for indexing)
-     *
-     * @return array
-     */
-    public function getAllProductAttributes()
-    {
-        if (is_null(self::$_productAttributes))
-        {
-            self::$_productAttributes = array();
-
-            /** @var $config Mage_Eav_Model_Config */
-            $config = Mage::getSingleton('eav/config');
-
-            $allAttributes = $config->getEntityAttributeCodes('catalog_product');
-
-            $productAttributes = array_merge(array('name', 'path', 'categories', 'description', 'ordered_qty', 'stock_qty', 'price_with_tax'), $allAttributes);
-
-            $excludedAttributes = array(
-                'all_children', 'available_sort_by', 'children', 'children_count', 'custom_apply_to_products',
-                'custom_design', 'custom_design_from', 'custom_design_to', 'custom_layout_update', 'custom_use_parent_settings',
-                'default_sort_by', 'display_mode', 'filter_price_range', 'global_position', 'image', 'include_in_menu', 'is_active',
-                'is_always_include_in_menu', 'is_anchor', 'landing_page', 'level', 'lower_cms_block',
-                'page_layout', 'path_in_store', 'position', 'small_image', 'thumbnail', 'url_key', 'url_path',
-                'visible_in_menu');
-
-            $productAttributes = array_diff($productAttributes, $excludedAttributes);
-
-            foreach ($productAttributes as $attributeCode)
-                self::$_productAttributes[$attributeCode] = $config->getAttribute('catalog_category', $attributeCode)->getFrontendLabel();
-        }
-
-        return self::$_productAttributes;
-    }
-
-    /**
-     * Return array of all category attributes that can be indexed (all except internal attributes and default attributes for indexing)
-     *
-     * @return array
-     */
-    public function getAllCategoryAttributes()
-    {
-        if (is_null(self::$_categoryAttributes))
-        {
-            self::$_categoryAttributes = array();
-
-            /** @var $config Mage_Eav_Model_Config */
-            $config = Mage::getSingleton('eav/config');
-
-            $allAttributes = $config->getEntityAttributeCodes('catalog_category');
-
-            $categoryAttributes = array_merge($allAttributes, array('product_count'));
-
-            $excludedAttributes = array(
-                'all_children', 'available_sort_by', 'children', 'children_count', 'custom_apply_to_products',
-                'custom_design', 'custom_design_from', 'custom_design_to', 'custom_layout_update', 'custom_use_parent_settings',
-                'default_sort_by', 'display_mode', 'filter_price_range', 'global_position', 'image', 'include_in_menu', 'is_active',
-                'is_always_include_in_menu', 'is_anchor', 'landing_page', 'level', 'lower_cms_block',
-                'page_layout', 'path_in_store', 'position', 'small_image', 'thumbnail', 'url_key', 'url_path',
-                'visible_in_menu');
-
-            $categoryAttributes = array_diff($categoryAttributes, $excludedAttributes);
-
-            foreach ($categoryAttributes as $attributeCode)
-                self::$_categoryAttributes[$attributeCode] = $config->getAttribute('catalog_category', $attributeCode)->getFrontendLabel();
-        }
-
-        return self::$_categoryAttributes;
-    }
-
-    /************/
-    /* Indexing */
-    /************/
-
-    /**
-     * Retrieve object id for the category
-     *
-     * @param Mage_Catalog_Model_Category $category
-     * @return string
-     */
-
-    private function try_cast($value)
-    {
-        if (is_numeric($value) && floatval($value) == floatval(intval($value)))
-            return intval($value);
-
-        if (is_numeric($value))
-            return floatval($value);
-
-        return $value;
-    }
-
-    private function getReportForProduct($product)
-    {
-        $report = Mage::getResourceModel('reports/product_collection')
-            ->addOrderedQty()
-            ->addAttributeToFilter('sku', $product->getSku())
-            ->setOrder('ordered_qty', 'desc')
-            ->getFirstItem();
-
-        return $report;
-    }
-
-    private function castProductObject(&$productData)
-    {
-        foreach ($productData as $key => &$data)
-        {
-            $data = $this->try_cast($data);
-
-            if (is_array($data) === false)
-            {
-                $data = explode('|', $data);
-
-                if (count($data) == 1)
-                {
-                    $data = $data[0];
-                    $data = $this->try_cast($data);
-                }
-                else
-                    foreach($data as &$element)
-                        $element = $this->try_cast($element);
-            }
-        }
-    }
-
-    private function isAttributeEnabled($additionalAttributes, $attr_name)
-    {
-        foreach ($additionalAttributes as $attr)
-            if ($attr['attribute'] == $attr_name)
-                return true;
-
-        return false;
-    }
-
-    public function removeProducts($ids, $store_id = NULL)
+    private function getStores($store_id)
     {
         $store_ids = array();
 
-        if ($store_id == NULL)
+        if ($store_id == null)
         {
             foreach (Mage::app()->getStores() as $store)
                 if ($store->getIsActive())
@@ -487,6 +88,13 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         }
         else
             $store_ids = array($store_id);
+
+        return $store_ids;
+    }
+
+    public function removeProducts($ids, $store_id = null)
+    {
+        $store_ids = $this->getStores($store_id);
 
         foreach ($store_ids as $store_id)
         {
@@ -496,18 +104,9 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         }
     }
 
-    public function removeCategories($ids, $store_id = NULL)
+    public function removeCategories($ids, $store_id = null)
     {
-        $store_ids = array();
-
-        if ($store_id == NULL)
-        {
-            foreach (Mage::app()->getStores() as $store)
-                if ($store->getIsActive())
-                    $store_ids[] = $store->getId();
-        }
-        else
-            $store_ids = array($store_id);
+        $store_ids = $this->getStores($store_id);
 
         foreach ($store_ids as $store_id)
         {
@@ -517,297 +116,22 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         }
     }
 
-    /**
-     * Prepare product JSON
-     *
-     * @param Mage_Catalog_Model_Product $product
-     * @param array                      $defaultData
-     * @return array
-     */
-    public function getProductJSON(Mage_Catalog_Model_Product $product, $defaultData = array())
-    {
-        $transport      = new Varien_Object($defaultData);
-
-        Mage::dispatchEvent('algolia_product_index_before', array('product' => $product, 'custom_data' => $transport));
-
-        $defaultData    = $transport->getData();
-
-        $defaultData    = is_array($defaultData) ? $defaultData : explode("|",$defaultData);
-
-        $customData = array(
-            'objectID'          => $product->getId(),
-            'name'              => $product->getName(),
-            'price'             => $product->getPrice(),
-            'price_with_tax'    => Mage::helper('tax')->getPrice($product, $product->getPrice(), true, null, null, null, null, false),
-            'url'               => $product->getProductUrl(),
-            'description'       => $product->getDescription()
-        );
-
-        $categories             = array();
-        $categories_with_path   = array();
-
-        foreach ($this->getProductActiveCategories($product, $product->getStoreId()) as $categoryId)
-        {
-            $category = Mage::getModel('catalog/category')->load($categoryId);
-
-            $categoryName = $category->getName();
-
-            if ($categoryName)
-                $categories[] = $categoryName;
-
-            $category->getUrlInstance()->setStore($product->getStoreId());
-            $path = '';
-
-            foreach ($category->getPathIds() as $treeCategoryId)
-            {
-                if ($path != '')
-                    $path .= ' /// ';
-
-                $path .= $this->getCategoryName($treeCategoryId, $product->getStoreId());
-            }
-
-            $categories_with_path[] = $path;
-        }
-
-        $customData['categories'] = $categories_with_path;
-
-        $customData['categories_without_path'] = $categories;
-
-        if (false === isset($defaultData['thumbnail_url']))
-        {
-            try
-            {
-                $customData['thumbnail_url'] = $product->getThumbnailUrl();
-                $customData['thumbnail_url'] = str_replace(array('https://', 'http://'
-                ), '//', $customData['thumbnail_url']);
-            }
-            catch (\Exception $e) {}
-        }
-
-        if (false === isset($defaultData['image_url']))
-        {
-            try
-            {
-                $customData['image_url'] = $product->getImageUrl();
-                $customData['image_url'] = str_replace(array('https://', 'http://'), '//', $customData['image_url']);
-            }
-            catch (\Exception $e) {}
-        }
-
-        $additionalAttributes = $this->getProductAdditionalAttributes($product->getStoreId());
-
-        // skip default calculation if we have provided these attributes via the observer in $defaultData
-        if (false === isset($defaultData['ordered_qty']) && false === isset($defaultData['stock_qty']))
-        {
-            $ordered_qty = Mage::getResourceModel('reports/product_collection')
-                ->addOrderedQty()
-                ->addAttributeToFilter('sku', $product->getSku())
-                ->setOrder('ordered_qty', 'desc')
-                ->getFirstItem()
-                ->ordered_qty;
-
-            $customData['ordered_qty'] = (int) $ordered_qty;
-            $customData['stock_qty']   = (int) Mage::getModel('cataloginventory/stock_item')->loadByProduct($product)->getQty();
-
-            if ($product->getTypeId() == 'configurable')
-            {
-                $sub_products = $product->getTypeInstance(true)->getUsedProducts(null, $product);
-                $ordered_qty  = 0;
-                $stock_qty    = 0;
-
-                foreach ($sub_products as $sub_product)
-                {
-                    $stock_qty += (int)Mage::getModel('cataloginventory/stock_item')->loadByProduct($sub_product)->getQty();
-
-                    $ordered_qty += (int) $this->getReportForProduct($sub_product)->ordered_qty;
-                }
-
-                $customData['ordered_qty'] = $ordered_qty;
-                $customData['stock_qty']   = $stock_qty;
-            }
-
-            if ($this->isAttributeEnabled($additionalAttributes, 'ordered_qty') == false)
-                unset($customData['ordered_qty']);
-
-            if ($this->isAttributeEnabled($additionalAttributes, 'stock_qty') == false)
-                unset($customData['stock_qty']);
-        }
-
-        foreach ($additionalAttributes as $attribute)
-        {
-            $value = $product->hasData($this->_dataPrefix . $attribute['attribute'])
-                ? $product->getData($this->_dataPrefix . $attribute['attribute'])
-                : $product->getData($attribute['attribute']);
-
-            $value = Mage::getResourceSingleton('algoliasearch/fulltext')->getAttributeValue($attribute['attribute'], $value, $product->getStoreId(), Mage_Catalog_Model_Product::ENTITY);
-
-            if ($value)
-                $customData[$attribute['attribute']] = $value;
-        }
-
-        $customData = array_merge($customData, $defaultData);
-
-        $customData['type_id'] = $product->getTypeId();
-
-        $this->castProductObject($customData);
-
-        return $customData;
-    }
-
-    /**
-     * Prepare category JSON
-     *
-     * @param Mage_Catalog_Model_Category $category
-     * @return array
-     */
-    public function getCategoryJSON(Mage_Catalog_Model_Category $category)
-    {
-        $transport = new Varien_Object();
-        Mage::dispatchEvent('algolia_category_index_before', array('category' => $category, 'custom_data' => $transport));
-        $customData = $transport->getData();
-
-        $storeId = $category->getStoreId();
-        $category->getUrlInstance()->setStore($storeId);
-        $path = '';
-        foreach ($category->getPathIds() as $categoryId) {
-            if ($path != '') {
-                $path .= ' / ';
-            }
-            $path .= $this->getCategoryName($categoryId, $storeId);
-        }
-
-        $image_url = NULL;
-        try {
-            $image_url = $category->getImageUrl();
-        } catch (Exception $e) { /* no image, no default: not fatal */
-        }
-        $data = array(
-            'objectID'      => $category->getId(),
-            'name'          => $category->getName(),
-            'path'          => $path,
-            'level'         => $category->getLevel(),
-            'url'           => $category->getUrl(),
-            '_tags'         => array('category'),
-            'popularity'    => 1,
-            'product_count' => $category->getProductCount()
-        );
-
-        if ( ! empty($image_url)) {
-            $data['image_url'] = $image_url;
-        }
-        foreach ($this->getCategoryAdditionalAttributes($storeId) as $attribute) {
-            $value = $category->hasData($this->_dataPrefix.$attribute['attribute'])
-                ? $category->getData($this->_dataPrefix.$attribute['attribute'])
-                : $category->getData($attribute['attribute']);
-
-            $value = Mage::getResourceSingleton('algoliasearch/fulltext')->getAttributeValue($attribute['attribute'], $value, $storeId, Mage_Catalog_Model_Category::ENTITY);
-
-            if (isset($data[$attribute['attribute']]))
-                $value = $data[$attribute['attribute']];
-
-            if ($value)
-                $data[$attribute['attribute']] = $value;
-        }
-
-        $data = array_merge($data, $customData);
-
-        foreach ($data as &$data0)
-            $data0 = $this->try_cast($data0);
-
-        return $data;
-    }
-
-    private function strip($s)
-    {
-        $s = trim(preg_replace('/\s+/', ' ', $s));
-        $s = preg_replace('/&nbsp;/', ' ', $s);
-        $s = preg_replace('!\s+!', ' ', $s);
-        return trim(strip_tags($s));
-    }
-
-    /**
-     * Adding product count when load collection is incorrect.
-     * The method applies the same limitation as on frontend to get correct product count for the category in the specified store.
-     * Product collection will not be loaded so this solution is fast.
-     *
-     * @param Mage_Catalog_Model_Category $category
-     * @return Algolia_Algoliasearch_Helper_Data
-     */
-    public function addCategoryProductCount(Mage_Catalog_Model_Category $category)
-    {
-        $productCollection = $category->getProductCollection(); /** @var $productCollection Mage_Catalog_Model_Resource_Product_Collection */
-        $category->setProductCount($productCollection->addMinimalPrice()->count());
-        return $this;
-    }
 
     public function rebuiltStorePageIndex($storeId)
     {
-        $emulationInfo = $this->startEmulation($storeId);
+        $index_name = $this->page_helper->getIndexName($storeId);
 
-        $indexer = $this->getIndex($this->getIndexName($storeId).'_pages_tmp');
+        $pages = $this->page_helper->getPages($storeId);
 
-        try
-        {
-            $magento_pages = Mage::getModel('cms/page')->getCollection()->addFieldToFilter('is_active',1);
+        foreach (array_chunk($pages, 100) as $chunk)
+            $this->algolia_helper->addObjects($chunk, $index_name.'_tmp');
 
-            $ids = $magento_pages->toOptionArray();
+        $this->algolia_helper->moveIndex($index_name.'_tmp', $index_name);
 
-            $excluded_pages = array_values($this->getExcludedPages());
-
-            foreach ($excluded_pages as &$excluded_page)
-                $excluded_page = $excluded_page['pages'];
-
-            $pages = array();
-
-            foreach ($ids as $key => $value)
-            {
-                if (in_array($value['value'], $excluded_pages))
-                    continue;
-
-                $page_obj = array();
-
-                $page_obj['slug'] = $value['value'];
-                $page_obj['name'] = $value['label'];
-
-                $page = Mage::getModel('cms/page');
-                $page->setStoreId($storeId);
-                $page->load($page_obj['slug'], 'identifier');
-
-                if (! $page->getId())
-                    continue;
-
-                $page_obj['objectID'] = $page->getId();
-
-                $page_obj['url'] = Mage::helper('cms/page')->getPageUrl($page->getId());
-                $page_obj['content'] = $this->strip($page->getContent());
-
-                $pages[] = $page_obj;
-            }
-
-            foreach (array_chunk($pages, 100) as $chunk)
-                $indexer->addObjects($chunk);
-
-            $this->getClient()->moveIndex($this->getIndexName($storeId).'_pages_tmp', $this->getIndexName($storeId).'_pages');
-        }
-        catch (\Exception $e)
-        {
-            $this->stopEmulation($emulationInfo);
-            throw $e;
-        }
-
-        $this->stopEmulation($emulationInfo);
+        $this->algolia_helper->setSettings($index_name, $this->page_helper->getIndexSettings($storeId));
     }
 
-
-    /**
-     * Rebuild store category index
-     *
-     * @param mixed          $storeId
-     * @param null|int|array $categoryIds
-     * @return void
-     * @throws Exception
-     */
-    public function rebuildStoreCategoryIndex($storeId, $categoryIds = NULL)
+    public function rebuildStoreCategoryIndex($storeId, $categoryIds = null)
     {
         $emulationInfo = $this->startEmulation($storeId);
 
@@ -819,7 +143,7 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 
             $categories = Mage::getResourceModel('catalog/category_collection'); /** @var $categories Mage_Catalog_Model_Resource_Eav_Mysql4_Category_Collection */
 
-            $unserializedCategorysAttrs = unserialize(Mage::getStoreConfig(self::XML_PATH_CATEGORY_ATTRIBUTES, $storeId));
+            $unserializedCategorysAttrs = $this->config->getCategoryAdditionalAttributes($storeId);
 
             $additionalAttr = array();
 
@@ -857,8 +181,6 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 
                         $category->setStoreId($storeId);
 
-                        $this->addCategoryProductCount($category);
-
                         $category_obj = $this->getCategoryJSON($category);
 
                         if ($category_obj['product_count'] > 0)
@@ -889,21 +211,12 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         $this->stopEmulation($emulationInfo);
     }
 
-    /**
-     * Rebuild store product index.
-     * Fallback to the default fulltext search indexer to prepare default data.
-     * After preparing default data, default data will be combined with custom data for Algolia search.
-     *
-     * @see Mage_CatalogSearch_Model_Resource_Fulltext::_rebuildStoreIndex()
-     * @see Mage_CatalogSearch_Model_Resource_Fulltext::_saveProductIndexes()
-     *
-     * @param mixed          $storeId
-     * @param null|int|array $productIds
-     * @param null|array     $defaultData
-     * @return void
-     * @throws Exception
-     */
-    public function rebuildStoreProductIndex($storeId, $productIds, $defaultData = NULL)
+    public function isEnabled($storeId = NULL)
+    {
+        return $this->config->isEnabled($storeId);
+    }
+
+    public function rebuildStoreProductIndex($storeId, $productIds, $defaultData = null)
     {
         if (count($productIds) > 1)
             $this->rebuiltStorePageIndex($storeId);
@@ -912,15 +225,15 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 
         try
         {
-            $indexer = $this->getIndex($this->getIndexName($storeId).'_products');
-            $products = Mage::getResourceModel('catalog/product_collection'); /** @var $products Mage_Catalog_Model_Resource_Eav_Mysql4_Product_Collection */
+            $index_name = $this->product_helper->getIndexName($storeId);
 
-            $additionalAttr = array();
+            /** @var $products Mage_Catalog_Model_Resource_Eav_Mysql4_Product_Collection */
+            $products = Mage::getResourceModel('catalog/product_collection');
 
-            $unserializedProductsAttrs = unserialize(Mage::getStoreConfig(self::XML_PATH_PRODUCT_ATTRIBUTES, $storeId));
+            $additionalAttr = $this->config->getProductAdditionalAttributes($storeId);
 
-            foreach ($unserializedProductsAttrs as $attr)
-                $additionalAttr[] = $attr['attribute'];
+            foreach ($additionalAttr as &$attr)
+                $attr = $attr['attribute'];
 
             $products
                 ->setStoreId($storeId)
@@ -957,13 +270,13 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 
                         $default = isset($defaultData[$product->getId()]) ? (array) $defaultData[$product->getId()] : array();
 
-                        $json = $this->getProductJSON($product, $default);
+                        $json = $this->product_helper->getObject($product, $default);
 
                         array_push($indexData, $json);
 
                         if (count($indexData) >= self::BATCH_SIZE)
                         {
-                            $indexer->addObjects($indexData);
+                            $this->algolia_helper->addObjects($indexData, $index_name);
                             $indexData = array();
                         }
                     }
@@ -975,7 +288,7 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
                 }
 
                 if (count($indexData) > 0)
-                    $indexer->addObjects($indexData);
+                    $this->algolia_helper->addObjects($indexData, $index_name);
 
                 unset($indexData);
             }
@@ -989,12 +302,6 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         $this->stopEmulation($emulationInfo);
     }
 
-    /**
-     * Start store emulation. Disable product and category flat catalog.
-     *
-     * @param mixed $storeId
-     * @return Varien_Object
-     */
     public function startEmulation($storeId)
     {
         $info = new Varien_Object;
@@ -1008,12 +315,6 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         return $info;
     }
 
-    /**
-     * Stop store emulation. Restore product and category flat catalog configuration.
-     *
-     * @param Varien_Object $info
-     * @return void
-     */
     public function stopEmulation($info)
     {
         Mage::app()->setCurrentStore($info->getInitialStoreId());
@@ -1025,166 +326,6 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
     /* Proxies */
     /***********/
 
-    /**
-     * Proxy for category names
-     *
-     * @param Mage_Catalog_Model_Category|int $categoryId
-     * @param Mage_Core_Model_Store|int $storeId
-     * @return null|string
-     */
-    public function getCategoryName($categoryId, $storeId = NULL)
-    {
-        if ($categoryId instanceof Mage_Catalog_Model_Category) {
-            $categoryId = $categoryId->getId();
-        }
-        if ($storeId instanceof Mage_Core_Model_Store) {
-            $storeId = $storeId->getId();
-        }
-        $categoryId = intval($categoryId);
-        $storeId = intval($storeId);
-
-        if (is_null(self::$_categoryNames)) {
-            self::$_categoryNames = array();
-            $resource = Mage::getResourceModel('catalog/category'); /** @var $resource Mage_Catalog_Model_Resource_Category */
-            if ($attribute = $resource->getAttribute('name')) {
-                $connection = Mage::getSingleton('core/resource')->getConnection('core_read'); /** @var $connection Varien_Db_Adapter_Pdo_Mysql */
-                $select = $connection->select()
-                    ->from(array('backend' => $attribute->getBackendTable()), array(new Zend_Db_Expr("CONCAT(backend.store_id, '-', backend.entity_id)"), 'backend.value'))
-                    ->join(array('category' => $resource->getTable('catalog/category')), 'backend.entity_id = category.entity_id', array())
-                    ->where('backend.entity_type_id = ?', $attribute->getEntityTypeId())
-                    ->where('backend.attribute_id = ?', $attribute->getAttributeId())
-                    ->where('category.level > ?', 1);
-                self::$_categoryNames = $connection->fetchPairs($select);
-            }
-        }
-
-        $categoryName = NULL;
-        $key = $storeId.'-'.$categoryId;
-        if (isset(self::$_categoryNames[$key])) { // Check whether the category name is present for the specified store
-            $categoryName = strval(self::$_categoryNames[$key]);
-        } elseif ($storeId != 0) { // Check whether the category name is present for the default store
-            $key = '0-'.$categoryId;
-            if (isset(self::$_categoryNames[$key])) {
-                $categoryName = strval(self::$_categoryNames[$key]);
-            }
-        }
-
-        return $categoryName;
-    }
-
-    /**
-     * Retrieve the list of all active categories
-     *
-     * @return array
-     */
-    public function getCategories()
-    {
-        if (is_null(self::$_activeCategories)) {
-            self::$_activeCategories = array();
-            $resource = Mage::getResourceModel('catalog/category'); /** @var $resource Mage_Catalog_Model_Resource_Category */
-            if ($attribute = $resource->getAttribute('is_active')) {
-                $connection = Mage::getSingleton('core/resource')->getConnection('core_read'); /** @var $connection Varien_Db_Adapter_Pdo_Mysql */
-                $select = $connection->select()
-                    ->from(array('backend' => $attribute->getBackendTable()), array('key' => new Zend_Db_Expr("CONCAT(backend.store_id, '-', backend.entity_id)"), 'category.path', 'backend.value'))
-                    ->join(array('category' => $resource->getTable('catalog/category')), 'backend.entity_id = category.entity_id', array())
-                    ->where('backend.entity_type_id = ?', $attribute->getEntityTypeId())
-                    ->where('backend.attribute_id = ?', $attribute->getAttributeId())
-                    ->order('backend.store_id')
-                    ->order('backend.entity_id');
-                self::$_activeCategories = $connection->fetchAssoc($select);
-            }
-        }
-        return self::$_activeCategories;
-    }
-
-    /**
-     * Retrieve category path.
-     * Category path can be found only for active categories.
-     *
-     * @param int $categoryId
-     * @param null|string $storeId
-     * @return null|string
-     */
-    public function getCategoryPath($categoryId, $storeId = NULL)
-    {
-        $categories = $this->getCategories();
-        $storeId = intval($storeId);
-        $categoryId = intval($categoryId);
-        $path = NULL;
-        $key = $storeId.'-'.$categoryId;
-        if (isset($categories[$key])) {
-            $path = ($categories[$key]['value'] == 1) ? strval($categories[$key]['path']) : NULL;
-        } elseif ($storeId !== 0) {
-            $key = '0-'.$categoryId;
-            if (isset($categories[$key])) {
-                $path = ($categories[$key]['value'] == 1) ? strval($categories[$key]['path']) : NULL;
-            }
-        }
-        return $path;
-    }
-
-    /**
-     * Check whether specified category is active
-     *
-     * @param int $categoryId
-     * @param null|int $storeId
-     * @return bool
-     */
-    public function isCategoryActive($categoryId, $storeId = NULL)
-    {
-        $storeId = intval($storeId);
-        $categoryId = intval($categoryId);
-        // Check whether the specified category is active
-        if ($path = $this->getCategoryPath($categoryId, $storeId)) {
-            // Check whether all parent categories for the current category are active
-            $isActive = TRUE;
-            $parentCategoryIds = explode('/', $path);
-            // Exclude root category
-            if (count($parentCategoryIds) <= 2) {
-                return FALSE;
-            }
-            // Remove root category
-            array_shift($parentCategoryIds);
-            // Remove current category as it is already verified
-            array_pop($parentCategoryIds);
-            // Start from the first parent
-            $parentCategoryIds = array_reverse($parentCategoryIds);
-            foreach ($parentCategoryIds as $parentCategoryId) {
-                if ( ! ($parentCategoryPath = $this->getCategoryPath($parentCategoryId, $storeId))) {
-                    $isActive = FALSE;
-                    break;
-                }
-            }
-            if ($isActive) {
-                return TRUE;
-            }
-        }
-        return FALSE;
-    }
-
-    /**
-     * Retrieve active categories for the product for the specified store
-     *
-     * @param int|Mage_Catalog_Model_Product $product
-     * @param int $storeId
-     * @return array
-     */
-    public function getProductActiveCategories(Mage_Catalog_Model_Product $product, $storeId = NULL)
-    {
-        $activeCategories = array();
-        foreach ($product->getCategoryIds() as $categoryId) {
-            if ($this->isCategoryActive($categoryId, $storeId)) {
-                $activeCategories[] = $categoryId;
-            }
-        }
-        return $activeCategories;
-    }
-
-    /**
-     * Retrieve root category id
-     *
-     * @return int
-     */
     public function getRootCategoryId()
     {
         if (-1 === self::$_rootCategoryId) {
@@ -1196,171 +337,4 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         }
         return self::$_rootCategoryId;
     }
-
-    /*************************/
-    /* Configuration getters */
-    /*************************/
-
-    public function getApplicationID($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_APPLICATION_ID, $storeId);
-    }
-
-    public function getAPIKey($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_API_KEY, $storeId);
-    }
-
-    public function getSearchOnlyAPIKey($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_SEARCH_ONLY_API_KEY, $storeId);
-    }
-
-    public function getIndexPrefix($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_INDEX_PREFIX, $storeId);
-    }
-
-    public function getCategoryAdditionalAttributes($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_CATEGORY_ATTRIBUTES, $storeId));
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getProductAdditionalAttributes($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_PRODUCT_ATTRIBUTES, $storeId));
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getFacets($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_FACETS, $storeId));
-
-        foreach ($attrs as &$attr)
-            if ($attr['type'] == 'other')
-                $attr['type'] = $attr['other_type'];
-
-        if (is_array($attrs))
-            return array_values($attrs);
-
-        return array();
-    }
-
-    public function getCategoryCustomRanking($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_CATEGORY_CUSTOM_RANKING, $storeId));
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getProductCustomRanking($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_PRODUCT_CUSTOM_RANKING, $storeId));
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getSaveLastQuery($storeId = NULL)
-    {
-        return Mage::getStoreConfigFlag(self::XML_PATH_SAVE_LAST_QUERY, $storeId);
-    }
-
-    public function isEnabled($storeId = NULL)
-    {
-        return (bool) Mage::getStoreConfigFlag(self::XML_PATH_IS_ALGOLIA_SEARCH_ENABLED, $storeId);
-    }
-
-    public function isPopupEnabled($storeId = NULL)
-    {
-        return ($this->isEnabled($storeId) && Mage::getStoreConfigFlag(self::XML_PATH_IS_POPUP_ENABLED, $storeId));
-    }
-
-    public function replaceCategories($storeId = NULL)
-    {
-        return Mage::getStoreConfigFlag(self::XML_PATH_REPLACE_CATEGORIES, $storeId);
-    }
-
-    public function isInstantEnabled($storeId = NULL)
-    {
-        return ($this->isEnabled($storeId) && Mage::getStoreConfigFlag(self::XML_PATH_IS_INSTANT_ENABLED, $storeId));
-    }
-
-    public function getInstantSelector($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_INSTANT_SELECTOR, $storeId);
-    }
-
-    public function getExcludedPages($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_EXCLUDED_PAGES, $storeId));
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getSortingIndices($storeId = NULL)
-    {
-        $attrs = unserialize(Mage::getStoreConfig(self::XML_PATH_SORTING_INDICES, $storeId));
-
-        foreach ($attrs as &$attr)
-            $attr['index_name'] = $this->getIndexName($storeId).'_products_'.$attr['attribute'].'_'.$attr['sort'];
-
-        if (is_array($attrs))
-            return $attrs;
-
-        return array();
-    }
-
-    public function getRemoveWordsIfNoResult($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_REMOVE_IF_NO_RESULT, $storeId);
-    }
-
-    public function getNumberOfProductSuggestions($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_NUMBER_OF_PRODUCT_SUGGESTIONS, $storeId);
-    }
-
-    public function getNumberOfProductResults($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_NUMBER_OF_PRODUCT_RESULTS, $storeId);
-    }
-
-    public function getNumberOfCategorySuggestions($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_NUMBER_OF_CATEGORY_SUGGESTIONS, $storeId);
-    }
-
-    public function getNumberOfPageSuggestions($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_NUMBER_OF_PAGE_SUGGESTIONS, $storeId);
-    }
-
-    public function getResultsLimit($storeId = NULL)
-    {
-        return Mage::getStoreConfig(self::XML_PATH_RESULTS_LIMIT, $storeId);
-    }
-
-    public function useResultCache($storeId = NULL)
-    {
-        return (bool) Mage::getStoreConfigFlag(self::XML_PATH_USE_RESULT_CACHE, $storeId);
-    }
-
 }
