+++ code/Model/Indexer/Algolia.php
@@ -82,9 +82,9 @@ class Algolia_Algoliasearch_Model_Indexer_Algolia extends Mage_Index_Model_Index
      */
     protected $_relatedConfigSettingsReindex = array(
         Mage_CatalogSearch_Model_Fulltext::XML_PATH_CATALOG_SEARCH_TYPE,
-        Algolia_Algoliasearch_Helper_Data::XML_PATH_CATEGORY_ATTRIBUTES,
-        Algolia_Algoliasearch_Helper_Data::XML_PATH_INDEX_PREFIX,
-        Algolia_Algoliasearch_Helper_Data::XML_PATH_INDEX_PRODUCT_COUNT
+        Algolia_Algoliasearch_Helper_Config::XML_PATH_CATEGORY_ATTRIBUTES,
+        Algolia_Algoliasearch_Helper_Config::XML_PATH_INDEX_PREFIX,
+        Algolia_Algoliasearch_Helper_Config::XML_PATH_INDEX_PRODUCT_COUNT
     );
 
     /**
