+++ code/Model/Resource/Fulltext/Collection.php
@@ -2,35 +2,31 @@
 
 class Algolia_Algoliasearch_Model_Resource_Fulltext_Collection extends Mage_CatalogSearch_Model_Resource_Fulltext_Collection
 {
-
     /**
-     * Add search query filter without preparing result since result table causes lots of lock contention.
-     *
-     * @param string $query
-     * @throws Exception
-     * @return Mage_CatalogSearch_Model_Resource_Fulltext_Collection
+     * Intercept query
      */
     public function addSearchFilter($query)
     {
-        if ( ! Mage::helper('algoliasearch')->isEnabled() || Mage::helper('algoliasearch')->useResultCache()) {
-            return parent::addSearchFilter($query);
+        $config = new Algolia_Algoliasearch_Helper_Config();
+
+        if ($config->isInstantEnabled())
+        {
+            $product_helper = new Algolia_Algoliasearch_Helper_Entity_Producthelper();
+
+            $url = Mage::getBaseUrl().'#q='.$query.'&page=0&refinements=%5B%5D&numerics_refinements=%7B%7D&index_name=%22'.$product_helper->getIndexName().'%22';
+
+            header('Location: '.$url);
+
+            die();
         }
 
-        // This method of filtering the product collection by the search result does not use the catalogsearch_result table
-        try {
-            $data = Mage::helper('algoliasearch')->getSearchResult($query, Mage::app()->getStore()->getId());
-        } catch (Exception $e) {
-            Mage::getSingleton('catalog/session')->addError(Mage::helper('algoliasearch')->__('Search failed. Please try again.'));
-            $this->getSelect()->columns(array('relevance' => new Zend_Db_Expr("e.entity_id")));
-            $this->getSelect()->where('e.entity_id = 0');
-            return $this;
-        }
+        $data = Mage::helper('algoliasearch')->getSearchResult($query, Mage::app()->getStore()->getId());
 
         $sortedIds = array_reverse(array_keys($data));
+
         $this->getSelect()->columns(array('relevance' => new Zend_Db_Expr("FIND_IN_SET(e.entity_id, '".implode(',',$sortedIds)."')")));
         $this->getSelect()->where('e.entity_id IN (?)', $sortedIds);
 
         return $this;
     }
-
 }
